requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.124.4
google-ai-generativelanguage==0.1.0
google-api-core==2.29.0
google-auth==2.49.2
google-generativeai==0.1.0rc1
googleapis-common-protos==1.73.0
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.9.1
openai==2.2.0
proto-plus==1.27.1
protobuf==4.25.9
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
sniffio==1.3.1
starlette==0.44.0
tqdm==4.67.3
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
